+++ epicshop/update-deps.sh
@@ -1,4 +1,4 @@
-npx npm-check-updates --dep prod,dev --upgrade --workspaces --root --reject msw
+npx npm-check-updates --dep prod,dev --upgrade --workspaces --root --reject eslint,@conform-to/react,@conform-to/zod
 cd epicshop && npx npm-check-updates --dep prod,dev --upgrade --root
 cd ..
 rm -rf node_modules package-lock.json ./epicshop/package-lock.json ./epicshop/node_modules ./exercises/**/node_modules