+++ exercises/04.unit-test/04.solution.setup/tests/setup/setup-test-env.ts
@@ -1,11 +1,11 @@
 import 'dotenv/config'
 import '#app/utils/env.server.ts'
 import { installGlobals } from '@remix-run/node'
-import { beforeEach, vi, type SpyInstance } from 'vitest'
+import { beforeEach, vi, type MockInstance } from 'vitest'
 
 installGlobals()
 
-export let consoleError: SpyInstance<Parameters<typeof console.error>>
+export let consoleError: MockInstance<typeof console.error>
 
 beforeEach(() => {
 	const originalConsoleError = console.error