+++ exercises/08.http-mocking/01.solution.start-server/tests/setup/setup-test-env.ts
@@ -3,13 +3,13 @@ import '#app/utils/env.server.ts'
 import '@testing-library/jest-dom/vitest'
 import { installGlobals } from '@remix-run/node'
 import { cleanup } from '@testing-library/react'
-import { afterEach, beforeEach, vi, type SpyInstance } from 'vitest'
+import { afterEach, beforeEach, vi, type MockInstance } from 'vitest'
 
 installGlobals()
 
 afterEach(() => cleanup())
 
-export let consoleError: SpyInstance<Parameters<typeof console.error>>
+export let consoleError: MockInstance<typeof console.error>
 
 beforeEach(() => {
 	const originalConsoleError = console.error